requirements.txt
httpx>=0.27
pydantic>=2.0
python-dotenv>=1.0
pyyaml>=6.0
fastapi>=0.110
uvicorn>=0.29
rich>=13.0
pypdf>=4.0
qdrant-client>=1.12
